refactor: API 레이어 DTO 및 E2E 테스트 수정

+++ apps/commerce-api/src/main/kotlin/com/loopers/interfaces/api/point/PointChargeDto.kt
@@ -2,14 +2,15 @@ package com.loopers.interfaces.api.point
 
 import com.loopers.application.point.PointInfo
 import jakarta.validation.constraints.NotBlank
+import java.math.BigDecimal
 
 class PointChargeDto {
     data class Request(
         @field:NotBlank(message = "포인트 충전량은 필수 입니다.")
-        val balance: Long,
+        val balance: BigDecimal,
     )
 
-    data class Response(val balance: Long) {
+    data class Response(val balance: BigDecimal) {
         companion object {
             fun from(info: PointInfo): Response = Response(
                 balance = info.balance,
